Finalizing alpha builder
