+++ docs/source/Builder.rst
@@ -1,0 +1,31 @@
+.. role:: cpp(code)
+   :language: c++
+
+.. _simba:
+
+======================================
+:fas:`keyboard;fa-fw` Builder (alpha)
+======================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+This is an Alamo input file builder for the **Integrator::Flame** method.
+Enter values for the Alamo parameters, and an input file is generated at the bottom of this page.
+Required parameters are indicated with red inputs.
+Default valuers are indicated with placeholder values, and will not be included in the input unless
+you put it in specifically.
+
+.. warning::
+   This is a new and very experimental Alamo feature.
+   This form is entirely auto-generated based on the Alamo code base using the code scrapers.
+   It is not currently complete (there are some known params that are not included)
+   and it is also not yet thoroughly tested.
+
+   Please use with caution.
+
+    
+.. raw:: html
+    :file: Builder.html
