--- docs/source/File.rst
@@ -1,15 +1,0 @@
-.. role:: cpp(code)
-   :language: c++
-
-.. _simba:
-
-======================================
-:fas:`question-circle;fa-fw` Questions
-======================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-.. raw:: html
-    :file: /home/brunnels/Research/alamo/index.html
