+++ docs/source/Inputs.py
@@ -7,6 +7,8 @@ from glob import glob
 
 from docutils.core import publish_string
 
+from scraper import getdocumentation, geticon, extract, scrapeInputs
+
 src2url = {}
 try:
 	doxsourcefiles = sorted(glob("../build/html/doxygen/*source.html"))
@@ -21,261 +23,6 @@ try:
 	                continue
 except Exception as e:
     print(e)
-#print(src2url)
-
-def geticon(classname):
-    if classname.startswith("BC"): return ":fas:`border-top-left;fa-fw` "
-    if classname.startswith("IC"): return ":fas:`circle-right;fa-fw` "
-    if classname.startswith("IO"): return ":fas:`print;fa-fw` "
-    if classname.startswith("Integrator"): return ":fas:`gear;fa-fw` "
-    if classname.startswith("Model"): return ":fas:`panorama;fa-fw` "
-    if classname.startswith("Numeric"): return ":fas:`calculator;fa-fw` "
-    if classname.startswith("Operator"): return ":far:`map;fa-fw` "
-    if classname.startswith("Set"): return ":fas:`braille;fa-fw` "
-    if classname.startswith("Solver"): return ":fas:`diamond-turn-right;fa-fw` "
-    if classname.startswith("Util"): return ":fas:`sliders;fa-fw` "
-    else: return ""
-
-def getdocumentation(filename):
-    sourcefile = open(filename+".H")
-    ret = ""
-    for line in sourcefile.readlines():
-        if line.startswith(r"///"): # special provision for legacy doxygen comments
-            ret += line.split(r"///")[1]
-        elif line.startswith(r"// "):
-            ret += line.split(r"// ")[1]
-        elif line.startswith(r"//"):
-            ret += line.split(r"//")[1]
-        else:
-            return ret
-    return ret
-
-def extract(basefilename):
-    rets = list()
-    class inputdoc: pass
-    for filename in [basefilename+".H",basefilename+".cpp"]:
-        if not os.path.isfile(filename): continue
-        sourcefile = open(filename)
-        lines = sourcefile.readlines()
-
-#        lines = clean_cpp_code(lines)
-#        for line in lines:
-#            print(line)
-
-        group = None
-        parsefn = False
-        switch = None
-        def reset():
-            nonlocal rets, group, parsefn, switch
-            if group:
-                rets.append(group)
-                group = None
-            if switch:
-                rets.append(switch)
-                switch = None
-        def insert(val):
-            nonlocal rets, group, parsefn, switch
-            val["parsefn"] = parsefn
-            if group: group["inputs"].append(val)
-            else: rets.append(val)
-
-        for i, line in enumerate(lines):
-            
-            # Catch standard pp.query and pp.queryarr inputs
-            match = re.findall('^\s*pp.(query[arr]*[_required]*[_file]*)\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',lines[i])
-            if match:
-                #print(match)
-                query = dict()
-                query["type"] = match[0][0]
-                query["string"] = match[0][1]
-                query["doc"] = match[0][2]
-                query["file"] = filename
-                query["line"] = i+1
-                
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    match = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if match: query["doc"] = match[0] + " " + query["doc"]
-                    else: break
-                insert(query)
-                continue
-
-            # Catch standard pp.query_default and pp.queryarr_default inputs
-            match = re.findall('^\s*pp.(query[arr]*_default*)\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,\s*"*([^"^,]+)"*\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',lines[i])
-            if match:
-                #print(match)
-                query = dict()
-                query["type"] = match[0][0]
-                query["string"] = match[0][1]
-                query["default"] = match[0][2]
-                query["doc"] = match[0][3]
-                query["file"] = filename
-                query["line"] = i+1
-                
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    match = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if match: query["doc"] = match[0] + " " + query["doc"]
-                    else: break
-                insert(query)
-                continue
-
-            # Catch standard pp.query_default and pp.queryarr_default inputs
-            match = re.findall('^\s*pp.(query_validate)\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,\s*\{(.*)\}\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',lines[i])
-            if match:
-                #print(match)
-                query = dict()
-                query["type"] = match[0][0]
-                query["string"] = match[0][1]
-                query["possibles"] = match[0][2]
-                query["doc"] = match[0][3]
-                query["file"] = filename
-                query["line"] = i+1
-                query["default"] = True
-                
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    match = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if match: query["doc"] = match[0] + " " + query["doc"]
-                    else: break
-                insert(query)
-                continue
-
-            # Catch pp.queryclass inputs
-            match = re.findall('^\s*pp.queryclass(?:<(.*)>)?\s*\(\s*"([^"]*)"(?:.*static_cast\s*<\s*(.*)\s*>.*)?[^)]*,*\s*[INFO]*\s*\);\s*(?:\/\/\s*(.*)$)?',line)
-            if match:
-                queryclass = dict()
-                queryclass["type"] = "queryclass"
-                queryclass["class"] = match[0][0]+match[0][2]
-                queryclass["string"] = match[0][1]
-                queryclass["doc"] = match[0][3]
-                queryclass["file"] = filename
-                queryclass["line"] = i+1
-
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    match = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if match: queryclass["doc"] = match[0] + " " + queryclass["doc"]
-                    else: break
-
-                insert(queryclass)
-                continue
-
-            # Catch ParmParse groups. This is old-fashioned but still supported.
-            # All subsequent inputs will be nested under a group
-            match = re.findall('^\s*(?:IO|amrex)::ParmParse\s*pp\s*(?:\(\s*"([^"]*)"\s*\))?\s*;\s*(?:\/\/\s*(.*))?',line)
-            if match:
-                reset()
-                group = dict()
-                group["type"] = "group"
-                group["prefix"] = match[0][0]
-                group["doc"] = match[0][1]
-                if group["doc"] == "":
-                    for j in reversed(range(0,i)):
-                        match = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                        if match: group["doc"] = match[0] + " " + group["doc"]
-                        else: break
-                group["inputs"] = list()
-
-            # Catch definition of a Parser function.
-            if re.match('^\s*(?!\/\/).*Parse\(.*(?:IO|amrex)::ParmParse\s*&\s*(pp)\)(?!\s*;)',line):
-                if parsefn: raise Exception(filename,i,"Multiple Parse functions cannot be declared in a single file")
-                parsefn = True
-
-
-            # Catch definition of a select function:
-            match = re.findall(r'pp\.select<([^>]+)>\s*\("([^"]+)"\s*,[^)]*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
-            if match:
-                input = dict()
-                input["type"] = "select"
-                input["classes"] = match[0][0].replace(' ','').split(',')
-                input["string"] = match[0][1].replace(' ','')
-                input["doc"] = match[0][2]
-
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    docmatch = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if docmatch:
-                        input["doc"] = docmatch[0] + " " + input["doc"]
-                    else: break
-                insert(input)
-
-            # Catch a queryclass 
-            match = re.findall(r'pp\.queryclass<([^>]+)>\s*\("([^"]+)"\s*,\s*[a-z,A-Z,0-9,_,.]*\s*,*\s*[INFO]*\s*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
-            if match:
-                input = dict()
-                input["type"] = "queryclass"
-                input["class"] = match[0][0].replace(' ','') 
-                input["string"] = match[0][1].replace(' ','')
-                input["doc"] = match[0][2]
-
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    docmatch = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if docmatch:
-                        input["doc"] = docmatch[0] + " " + input["doc"]
-                    else: break
-                insert(input)
-                continue
-
-            # Catch a queryclass with no ID
-            match = re.findall(r'pp\.queryclass<([^>]+)>\s*\([^\)]*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
-            match = re.findall(r'pp\.queryclass<([^\)_]+)>\s*\([^\)]*\)\s*;\s*(?:\/\/\s*(.*))?$',line)
-            if match:
-                input = dict()
-                input["type"] = "querysubclass"
-                input["class"] = match[0][0].replace(' ','') 
-                input["doc"] = match[0][1]
-                print(match)
-
-                # Check if previous lines have simple comments. Ignores "///" comments and
-                # any comment beginning with [
-                for j in reversed(range(0,i)):
-                    docmatch = re.findall('^\s*\/\/(?!\/)(?!\s*\[)\s*(.*)',lines[j])
-                    if docmatch:
-                        input["doc"] = docmatch[0] + " " + input["doc"]
-                    else: break
-                insert(input)
-            
-
-
-            # Catch definition of a switch group
-            match =re.findall('^\s*\/\/\s*\[\s*switch\s*(\S*)\s*]\s*(.*)',line)
-            if match:
-                reset()
-                switch = dict()
-                switch["type"] = "switch"
-                switch["string"] = match[0][0]
-                switch["doc"] = match[0][1]
-                switch["inputs"] = list()
-                continue
-
-            if switch:
-                # If we are inside a switch block, look for matches like
-                #    else if (ic_type == "KEY") value.ic = new CLASS(args,"PREFIX")
-                match = re.findall('^\s*(?:else)*\s*if\s*\(\S*\s*==\s*"([^"]*)"\s*\).*new\s*([A-Z,:,a-z,0-9,_]*)[^"]*"(.*)"',line)
-                if match:
-                    input=dict()
-                    input["type"]="switchitem"
-                    input["string"]=match[0][0]
-                    input["class"]=match[0][1]
-                    input["prefix"]=match[0][2]
-                    switch["inputs"].append(input)
-
-            # Close a switch group
-            if re.match('^\s*\/\/\s*\[\s*end\s*(?:switch)?\s*]',line):
-                reset()
-            
-        reset()
-
-    return rets
-
 
 inputsheader = r"""
 .. _inputs: 
@@ -368,377 +115,6 @@ the prefixes out.
       - Namespace / Class
       - Description
 """
-
-
-
-
-def scrapeInputsSimple(root="../../src/", writeFiles=True):
-
-    srcfiles = set()
-    for dirname, subdirlist, filelist in sorted(os.walk(root)):
-        for f in filelist:
-            if f.endswith(".cpp"): srcfiles.add(dirname+"/"+f.replace(".cpp",""))
-            if f.endswith(".H"): srcfiles.add(dirname+"/"+f.replace(".H",""))
-    srcfiles = list(srcfiles)
-    
-    headerchar = ["=","*","-","~","."]
-    written_headers = []
-    
-    global num_tot, num_doc
-    num_tot = 0
-    num_doc = 0
-
-    data=dict()
-
-    for dirname, subdirlist, filelist in sorted(os.walk(root)):
-        hdrname = dirname.replace(root,"").replace("/","::")
-        depth = len(hdrname.split("::")) 
-    
-        srcfileset = set()
-        for f in filelist:
-            if f.endswith(".cpp"): srcfileset.add(f.replace(".cpp",""))
-            if f.endswith(".H"): srcfileset.add(f.replace(".H",""))
-        srcfilelist = list(srcfileset)
-        
-        for f in sorted(srcfilelist):
-            classname = dirname.replace(root,"").replace("/","::") + "::" + f.replace(".H","").replace(".cpp","")
-            data[classname] = dict()
-
-            try:
-                data[classname]['inputs'] = extract(dirname+"/"+f)
-            except Exception as e:
-                print("ERROR: problem reading",dirname)
-                raise
-            data[classname]['documentation'] = getdocumentation(dirname+"/"+f)
-
-    def printInputs(classname,indent="",prefix=""):
-        
-        if classname not in data.keys(): return ""
-
-        html = ""
-
-        for input in data[classname]["inputs"]:
-            if input["type"] == "group": continue
-
-
-            if (input["type"] == "select"):
-                id = prefix + input["string"]
-
-                html     += indent + """<div class="card mb-3" >\n"""
-                html     += indent + """  <div class="card-header">\n"""
-                html     += indent + """    <h5>{}</h5>\n""".format(id + ".type")
-                html     += indent + """    <ul class="nav nav-tabs card-header-tabs" id="myTab" role="tablist">\n"""
-                for cls in input["classes"]:
-                    name = cls.split("::")[-1].lower()
-                    html += indent + """      <li class="nav-item">\n"""
-                    html += indent + """        <a class="nav-link" id="profile-tab" data-bs-toggle="tab" href="#{}" role="tab" >{}</a> \n""".format(input["string"] + "." + name, name)
-                    html += indent + """      </li>\n"""
-                html +=     indent + """    </ul>\n"""
-                html +=     indent + """  </div>\n"""
-                html +=     indent + """  <div class="card-body">\n"""
-                html +=     indent + """    <div class="tab-content">\n"""
-                for cls in input["classes"]:
-                    name = cls.split("::")[-1].lower()
-                    html += indent + """      <div class="tab-pane" id="{}" role="tabpanel">""".format(input["string"] + "." + name)
-
-
-                    html += indent + "<p>"
-                    html += data[cls]["documentation"]
-                    html += indent + "</p>"
-
-
-                    html += printInputs(cls,indent+"      ",
-                                        prefix = input["string"] + "." + name + ".")
-                    html += indent + """      </div>\n"""
-                html +=     indent + """    </div>\n"""
-                html +=     indent + """  </div>\n"""
-                html +=     indent + """</div>\n"""
-
-                html     += indent + """<input type='hidden' name='{}' value='[not completed yet - please enter manually]' >\n""".format(id+".type")
-
-
-            elif (input["type"] == "queryclass"):
-                id = prefix + input["string"]
-
-
-                html     += indent + """<div class="card mb-3" >\n"""
-                html     += indent + """  <div class="card-header">\n"""
-                html     += indent + """    <h5>{}</h5>\n""".format(id)
-                html +=     indent + """    </ul>\n"""
-                html +=     indent + """  </div>\n"""
-                html +=     indent + """  <div class="card-body">\n"""
-                html +=     indent + """    <div class="tab-content">\n"""
-
-                cls = input["class"]
-                cls = cls.split("<")[0] # remove template args for now
-                if not cls in data.keys(): # if it's not an actual classname then
-                    cls = "::".join(classname.split("::")[:-1])+"::"+cls
-                print(cls)
-                name = cls.split("::")[-1].lower()
-                
-                html += printInputs(cls,indent+"      ",
-                                    prefix = input["string"] + ".")
-
-                html +=     indent + """    </div>\n"""
-                html +=     indent + """  </div>\n"""
-                html +=     indent + """</div>\n"""
-
-            elif (input["type"] == "querysubclass"):
-                cls = input["class"]
-                cls = cls.split("<")[0] # remove template args for now
-                if not cls in data.keys(): # if it's not an actual classname then
-                    cls = "::".join(classname.split("::")[:-1])+"::"+cls
-                print(cls)
-
-                html += printInputs(cls,indent=indent, prefix=prefix)
-
-
-
-            elif "validate" in input["type"]:
-                id = prefix + input["string"]
-
-                html += '<div class="input-group mb-3">'
-                html += '  <div class="input-group-prepend">'
-                html += indent + "    <span class='input-group-text'> {} </span>\n".format(id)
-                html += indent + "  </div>\n"
-                html += indent + """  <select class="form-select" id="{}">""".format(id)
-                html += indent + """  <option selected disabled value="">Choose...</option>"""
-                print(input["possibles"].split(','))
-                for option in input["possibles"].replace('"','').split(','):
-                    html += indent + """  <option>{}</option>""".format(option)
-                html += indent + """  </select>"""
-
-                html += indent + "</div>"
-                
-            else:
-                id = prefix + input["string"]
-
-                html += '<div class="input-group mb-3">'
-                html += '  <div class="input-group-prepend">'
-                html += indent + """<span type="button" class="input-group-text" data-bs-toggle="popover" data-bs-title="{}" data-bs-content="{}">{}</span>""".format(input["file"].split("/")[-1]+":"+str(input["line"])
-                                                                                                                                                                      ,input["doc"],id)
-                #html += indent + "    <a class='input-group-text' data-bs-toggle='popover' data-content='documentation'> {} </a>\n".format(id)
-                html += indent + "  </div>\n"
-                required = ""
-                if 'required' in input["type"]: required="required"
-                placeholder = ""
-                if 'default' in input["type"]: placeholder=input["default"]
-                html += indent + "  <input type='text' id='{}' name='{}' class='form-control' {} placeholder='{}'>\n".format(
-                    id,id,required,placeholder)
-                html += indent + "</div>"
-
-        return html
-
-
-    html = """
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dynamic Form with Nested Sections</title>
-    <!-- Bootstrap CSS -->
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-
-</head>
-<body>
-
-
-<style>
-        .code-container {
-            position: relative;
-            background-color: #f8f9fa;
-            border: 1px solid #ddd;
-            border-radius: 5px;
-            padding: 10px;
-            font-family: monospace;
-            white-space: pre-wrap;
-            overflow-x: auto;
-        }
-        .variable { color: #007bff; } /* Blue for variables */
-        .string { color: #28a745; }   /* Green for strings */
-        .comment { color: #6c757d; }  /* Grey for comments */
-        .copy-button {
-            position: absolute;
-            top: 10px;
-            right: 10px;
-        }
-</style>
-
-
-
-
-<div class='container mt-5 mb-5'> 
-<h2 class='mb-4'>Settings Form</h2><form  id='form'>
-
-"""
-    #html += printInputs("Integrator::PhaseFieldMicrostructure")
-    #html += printInputs("Integrator::ThermoElastic")
-    html += printInputs("Integrator::Flame")
-    html += r"""
-</form>
-
-
-<form>
-  <h2 class='mb-4'>Alamo input file</h2><form  id='form'>
-  <div class="form-group">
-    <label for="code-container">The following contains only set values and is updated dynamically.</label>
-    <textarea class="form-control" id="code-container" rows="10"></textarea>
-  </div>
-</form>
-
-
-</div>
-</body>
-
-
-
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"></script>
-<script>
-
-
-
-function updateResult() {
-    // Select all input elements within the form (except the result textbox)
-    const form = document.getElementById('form');
-    const inputs = form.querySelectorAll('input, select');
-    //console.log(inputs);
-
-    // Concatenate the values of all inputs
-    //let concatenatedString = Array.from(inputs)
-    //    .map(input => input.value.trim()) // Get trimmed values
-    //    .join(" ");                      // Concatenate with a space
-
-    var concatenatedString = ""
-    inputs.forEach((input) => {
-      if(input.value.trim() != '')
-          concatenatedString += input.name + " = " + input.value + "\n";
-    });
-
-
-    // Update the result textbox
-    //document.getElementById("result").value = concatenatedString;
-    //console.log(concatenatedString);
-    document.getElementById("code-container").value = concatenatedString;
-}
-
-
-// Add event listeners to all inputs within the form (except the result textbox)
-document.querySelectorAll("input, select").forEach(input => {
-    input.addEventListener("input", updateResult); // Trigger on any input change
-});
-
-
-//    function generateInputFile() {
-//        const form = document.getElementById('form');
-//        const inputs = form.querySelectorAll('input');
-//
-//        inputs.forEach((input) => {
-//            if(input.value.trim() != '')
-//            console.log(input.name + "=" + input.value);
-//        });
-//        console.log(form)
-//        console.log(inputs)
-//        console.log(inputs.size)
-//        for (i=0; i < inputs.length; i++) {
-//           console.log(inputs.entries[i].id);
-//           console.log(inputs.entries[i].value);
-//        }
-//    }
-
-
-  // JavaScript for live validation
-  document.addEventListener('DOMContentLoaded', function () {
-    const form = document.getElementById('form');
-    const inputs = form.querySelectorAll('input[required]');
-
-
-    // Initialize all popovers
-    inputs.forEach((input) => {
-      new bootstrap.Popover(input);
-    });
-
-
-    // Validate fields on page load
-    inputs.forEach((input) => {
-      if (input.value.trim() === '') {
-        input.classList.add('is-invalid');
-      }
-    });
-
-    inputs.forEach((input) => {
-      input.addEventListener('input', function () {
-        if (!input.value) {
-          input.classList.add('is-invalid');
-        } 
-        else if (input.value.trim() === '') {
-          input.classList.add('is-invalid');
-        } else {
-          input.classList.remove('is-invalid');
-          input.classList.add('is-valid');
-        }
-      });
-    });
-
-      form.addEventListener('submit', function (event) {
-        let isValid = true;
-
-        inputs.forEach((input) => {
-          if (input.value.trim() === '') {
-            input.classList.add('is-invalid');
-            isValid = false;
-          } else {
-            input.classList.remove('is-invalid');
-          }
-        });
-
-      if (!isValid) {
-        event.preventDefault();
-        event.stopPropagation();
-      }
-    });
-
-
-    // Add event listener for all tabs across all containers
-    document.querySelectorAll('.tab-container').forEach(container => {
-        const hiddenInput = container.querySelector('.selectedTab');
-        const tabs = container.querySelectorAll('.nav-tabs .nav-link');
-
-        tabs.forEach(tab => {
-            tab.addEventListener('shown.bs.tab', function (event) {
-                const activeTabId = event.target.id; // Get the ID of the active tab
-                hiddenInput.value = activeTabId; // Update the hidden input value
-                console.log(activeTabId);
-            });
-        });
-    });
-  });
-
-
-const popoverTriggerList = document.querySelectorAll('[data-bs-toggle="popover"]')
-const popoverList = [...popoverTriggerList].map(popoverTriggerEl => new bootstrap.Popover(popoverTriggerEl))
-
-</script>
-
-
-"""
-    
-
-
-
-
-    f = open("index.html",'w')
-    f.write(html)
-    f.close()
-    #print(html)
-
-
-    
-
-
 
 def scrapeInputs(root="../../src/", writeFiles=True):
