+++ docs/source/conf.py
@@ -24,7 +24,7 @@ sys.path.append(os.path.abspath('.'))
 import Inputs
 Inputs.scrapeInputs()
 import Tests
-
+import Builder
 # -- Project information ----------------------------------------------------
