+++ docs/source/index.rst
@@ -170,4 +170,4 @@ The development team gratefully acknowledges the funding sources that enable the
         InputsSearch
         Developers
         Questions
-        File
+        Builder
